REL-403: moved TargetDesc and ObsTargetDesc to make skycalc available to TargetDesc
