+++ bundle/jsky.app.ot/src/main/java/jsky/app/ot/gemini/obscat/ObsCatalogQueryResultDisplay.java
@@ -5,7 +5,7 @@ import edu.gemini.pot.sp.*;
 import edu.gemini.pot.spdb.IDBDatabaseService;
 import edu.gemini.spModel.core.SPProgramID;
 import edu.gemini.spModel.core.Site;
-import edu.gemini.spModel.target.ObsTargetDesc;
+import jsky.app.ot.plot.ObsTargetDesc;
 import jsky.app.ot.OTOptions;
 import jsky.app.ot.session.SessionQueue;
 import jsky.app.ot.session.SessionQueuePanel;
@@ -22,7 +22,7 @@ import jsky.catalog.TableQueryResult;
 import jsky.catalog.gui.QueryResultDisplay;
 import jsky.catalog.gui.TableDisplay;
 import jsky.catalog.gui.TableDisplayTool;
-import jsky.coords.TargetDesc;
+import jsky.plot.TargetDesc;
 import jsky.plot.ElevationPlotManager;
 import jsky.util.gui.DialogUtil;
 import jsky.util.gui.SortedJTable;
