+++ bundle/jsky.elevation.plot/src/main/java/jsky/plot/ObservationPanel.java
@@ -23,7 +23,6 @@ import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTabbedPane;
 
-import jsky.coords.TargetDesc;
 import jsky.util.I18N;
 import jsky.util.PrintableWithDialog;
 import jsky.util.SaveableWithDialog;
